Client Storing

+++ build.sh
@@ -24,15 +24,6 @@ then
     cd build
     ./executable server
   fi
-elif [ "$1" = "run" ]  && [ "$2" = "client" ]
-then
-  if ! [ -d "./build" ]
-  then
-    echo "Build folder not located: ./build.sh build"
-  else
-    cd build
-    ./executable client 
-  fi
 elif [ "$1" = "test" ]
 then
   if ! [ -d "./build" ]
@@ -42,7 +33,7 @@ then
     cd build
     ./test_runner
   fi
-elif [ "$1" = "build" ] && [ "$2" = "run" ]
+elif [ "$1" = "build" ] && [ "$2" = "run" ] && [ "$3" = "" ]
 then
   rm -rf build/res
   cp -r res build/res
@@ -50,6 +41,14 @@ then
   cmake .. -DCMAKE_C_COMPILER=gcc -DCMAKE_EXPORT_COMPILE_COMMANDS=1
   make
   ./executable
+elif [ "$1" = "build" ] && [ "$2" = "run" ] && [ "$3" = "server" ]
+then
+  rm -rf build/res
+  cp -r res build/res
+  cd build
+  cmake .. -DCMAKE_C_COMPILER=gcc -DCMAKE_EXPORT_COMPILE_COMMANDS=1
+  make
+  ./executable server
 elif [ "$1" = "build" ] && [ "$2" = "test" ]
 then
   rm -rf build/res
@@ -58,7 +57,7 @@ then
   cmake .. -DCMAKE_C_COMPILER=gcc -DCMAKE_EXPORT_COMPILE_COMMANDS=1
   cmake --build .
   ./test_runner
-elif [ "$1" = "build" ] && [ "$2" = "test" ] && [ "$3" = "run" ]
+elif [ "$1" = "build" ] && [ "$2" = "test" ] && [ "$3" = "run" ] && [ "$4" = "" ]
 then
   rm -rf build/res
   cp -r res build/res
@@ -67,6 +66,15 @@ then
   cmake --build .
   ./test_runner
   ./executable
+elif [ "$1" = "build" ] && [ "$2" = "test" ] && [ "$3" = "run" ] && [ "$4" = "server" ]
+then
+  rm -rf build/res
+  cp -r res build/res
+  cd build
+  cmake .. -DCMAKE_C_COMPILER=gcc -DCMAKE_EXPORT_COMPILE_COMMANDS=1
+  cmake --build .
+  ./test_runner
+  ./executable server
 else
-  echo "Command usage: ./build.sh command [command: build, run, build run, test]"
+  echo "Command keywords: [build, run, server, test]"
 fi
